// types/enums.ts
export enum ProductCategory {
  Electronics = "electronics",
  Clothing = "clothing",
  Accessories = "accessories",
  Footwear = "footwear",
  Home = "home",
}

export enum ProductStatus {
  InStock = "in_stock",
  OutOfStock = "out_of_stock",
  PreOrder = "pre_order",
  LimitedEdition = "limited_edition",
}

export enum ProductSize {
  XS = "XS",
  S = "S",
  M = "M",
  L = "L",
  XL = "XL",
  XXL = "XXL",
  OneSize = "ONE SIZE",
}

export enum ProductColor {
  Black = "black",
  White = "white",
  Gray = "gray",
  Navy = "navy",
  Red = "red",
  Green = "green",
  Blue = "blue",
  Yellow = "yellow",
  Orange = "orange",
  Purple = "purple",
  Pink = "pink",
  Brown = "brown",
  Beige = "beige",
  Olive = "olive",
  Teal = "teal",
}

export enum SortOption {
  Newest = "newest",
  PriceLowToHigh = "price_low_to_high",
  PriceHighToLow = "price_high_to_low",
  Popular = "popular",
  Rating = "rating",
}

export enum ToastType {
  Success = "success",
  Error = "error",
  Warning = "warning",
  Info = "info",
}